\name{semsfa-package}
\alias{semsfa-package}
\docType{package}
\title{Semiparametric Stochastic Frontier Models}
\description{
Semiparametric Estimation of Stochastic Frontier Models following the two step procedure originally proposed by Fan et al (1996) and further developed also by Vidoli and Ferrara (2014) and Ferrara and Vidoli (2017). In the first step semiparametric or nonparametric regression techniques are used to relax parametric restrictions regards the functional form representing technology and in the second step variance parameters are obtained by pseudolikelihood or method of moment estimators.}

\author{
Giancarlo Ferrara, Francesco Vidoli\cr
Maintainer: Giancarlo Ferrara <giancarlo.ferrara@gmail.com>}

\references{
Aigner., D., Lovell, C.A.K., Schmidt, P., 1977. Formulation and estimation of stochastic frontier production function models. Journal of Econometrics 6:21-37

Fan, Y., Li, Q., Weersink, A., 1996. Semiparametric estimation of stochastic production frontier models. Journal of Business & Economic Statistics 14:460-468

Ferrara G., Vidoli F. (2017). Semiparametric stochastic frontier models: A generalized additive model approach. European Journal of Operational Research 258:761-777.

Hastie, T., Tibshirani, R., 1990. Generalized additive models. Chapman & Hall

Kumbhakar, S.C., Lovell, C.A.K, 2000. Stochastic Frontier Analysis. Cambridge University
Press, U.K

Meeusen, W., van den Broeck, J., 1977. Efficiency estimation from Cobb-Douglas production functions with composed error. International Economic Review 18:435-444

Vidoli, F., Ferrara, G., 2014. Analyzing Italian citrus sector by semi-nonparametric frontier efficiency models. Empirical Economics doi 10.1007/s00181-014-0879-6}